# contracts/openapi/salas-partidas.yaml
openapi: 3.1.0
...
info:
  title: Nexus Battles VI — Salas y partidas (salas-partidas)
  description: |
    API de creacion, listado e ingreso a salas de batalla, y de consulta del
    estado de una partida en curso.

    El avance del combate turno a turno NO viaja por esta API: viaja por el
    canal WebSocket descrito en `contracts/websocket/salas-partidas.yaml`.
    Esta API cubre el ciclo de vida de la sala hasta el inicio del combate, y
    la consulta puntual de estado para reconexion.

    Errores: todos los errores se devuelven en formato problem details
    (RFC 7807) con `Content-Type: application/problem+json`, identico en los 20
    modulos (regla 4 de plataforma). La correspondencia entre este formato y el
    componente `Aviso` de la interfaz esta en
    `shared/ui-kit/MAPEO-ERRORES.md`.

    Trazas: toda peticion propaga `traceparent` (regla 5 de plataforma).

    Cambios 1.2.0 (HU-SAL-004, aditivos): `heroesIA` en `CrearSalaRequest` y en
    `Sala` — cuantos cupos son de la inteligencia artificial; cuentan en
    `ocupacion`. `CONTRA_IA` lleva siempre uno; `UNO_CONTRA_UNO` ninguno;
    `HASTA_SEIS` de cero a `maximoParticipantes - 1`. `incluirHeroeIA` se
    conserva y vale por uno. En el modo cooperativo (`tamanoEquipo`) los
    participantes de la partida traen `equipo` asignado por orden de entrada.

    Cambios 1.3.0 (HU-DIS-003, degradacion controlada): cuando una dependencia
    de este servicio NO RESPONDE (inventario, motor de combate, libro de
...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /partidas/{idPartida}:
    parameters:
      - name: idPartida
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Partidas]
      summary: Consultar el estado de una partida
      description: |
        Devuelve el estado completo del combate: participantes con su vida,
        turno en curso y efectos activos.

        Se consulta al entrar al combate y al reconectar tras una caida del
        canal en tiempo real. Durante el juego normal el estado llega por
        WebSocket, no por esta ruta.

        Desde 1.7.0 (B7) solo la ven sus participantes y los roles de
        operacion (MODERADOR, ADMINISTRADOR, SUPER_ADMINISTRADOR, que ya
        podian leer partidas para atender reportes). Un jugador que no juega
        esta partida recibe `403` con `type` `partida-ajena`: el poder, las
        cargas y las acciones disponibles de cada heroe son informacion de la
        partida, no publica.
      operationId: obtenerPartida
      responses:
        '200':
          description: Estado de la partida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partida'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          description: El jugador no participa en esta partida (1.7.0).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
              example:
                type: 'https://nexusbattles.local/errores/partida-ajena'
                title: 'No participas en esta partida'
                status: 403
                detail: 'Solo quienes juegan una partida pueden ver su estado.'
        '404':
          description: Partida no encontrada
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  # ---------------------------------------------------------------------------
  # 1.10.0 · Revision del modo jugador (6-oct, punto 19) — Salir = rendirse
  # ---------------------------------------------------------------------------
  /partidas/{idPartida}/rendicion:
    parameters:
      - name: idPartida
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Partidas]
      summary: Rendirse (salir del combate cuenta como derrota)
      description: |
        1.10.0. Quien se rinde sale del token, nunca del cuerpo; solo un
        participante humano de la partida puede hacerlo (rol JUGADOR).

        El heroe de quien se rinde queda fuera de combate, como si hubiera
        caido; no se inventa un golpe ni un ejecutor. A partir de ahi manda
        la regla de siempre:

        - Si su bando se queda sin nadie en pie, la partida termina: gana
          quien sigue en pie, la apuesta se liquida con las reglas de la
          partida (las mismas que con el ultimo golpe) y se acredita lo de
          jugar. Llega `partida.finalizada` por el canal.
        - Si no termina y era su turno, el turno pasa (y juega la maquina si
          le toca), con motivo `RENDICION` en `partida.turno.cambiado`.
        - Si no era su turno, el combate sigue sin el.

        Antes que todo eso, a todos les llega `partida.participante.rendido`
        (AsyncAPI). Idempotente: con la partida terminada o el heroe ya
        fuera de combate devuelve la partida tal cual y no anuncia nada.
      operationId: rendirse
      responses:
        '200':
          description: La partida despues de la rendicion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partida'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          description: |
            `partida-ajena`: quien llama no combate en esta partida. Un rol
            de operacion no se rinde por nadie.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '404':
          description: Partida no encontrada
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '409':
          description: |
            `accion-no-permitida` con motivo `RENDICION_NO_PERMITIDA`: no se
            conoce el heroe del participante (partidas anteriores a las fichas).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  # ---------------------------------------------------------------------------
  # 1.7.0 (B7) — Historial de partidas del jugador
  # ---------------------------------------------------------------------------
  /partidas/mias:
    get:
      tags: [Partidas]
      summary: Historial de partidas del jugador autenticado (1.7.0)
      description: |
        Las partidas en las que jugo quien firma el token (el jugador sale del
        `uid`, nunca de la ruta), de la mas reciente a la mas antigua: en
        curso y terminadas, con su resultado desde su punto de vista. Mismo
        tamano de pagina que el listado de salas (RNF-USA-001: dieciseis).
      operationId: misPartidas
      parameters:
        - name: pagina
          in: query
          description: Numero de pagina, base 0
          schema: { type: integer, minimum: 0, default: 0 }
        - name: tamano
          in: query
          description: Elementos por pagina. El valor de diseno es 16.
          schema: { type: integer, minimum: 1, maximum: 50, default: 16 }
      responses:
        '200':
          description: Pagina del historial
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaDePartidas'
        '401':
          $ref: '#/components/responses/NoAutenticado'
...
components:
...
  responses:
...
    NoAutenticado:
      description: Falta el token o esta caducado
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    # HU-DIS-003 — la misma forma en toda operacion que necesita a otro servicio.
...
  schemas:
...
    ProblemDetail:
      type: object
      description: |
        Problem details RFC 7807. Identico en los 20 modulos.
        Correspondencia con la interfaz en `shared/ui-kit/MAPEO-ERRORES.md`.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Identificador estable del tipo de error. La interfaz decide por aqui, nunca por el texto.
        title:
          type: string
          description: Titulo corto. Va al titulo del componente Aviso.
        status:
          type: integer
          description: Codigo HTTP. Determina el tono del Aviso.
        detail:
          type: string
          description: Explicacion legible para el jugador. Va al cuerpo del Aviso.
        instance:
          type: string
          description: Ruta que produjo el error.
        errores:
          type: array
          description: Errores por campo, presente solo en validaciones de formulario.
          items:
            type: object
            required: [campo, mensaje]
            properties:
              campo: { type: string }
              mensaje: { type: string }

    # -- Enumeraciones ---------------------------------------------------------
    Modalidad:
      type: string
      description: |
        Modalidades admitidas (RF-JUE-004): uno contra uno, uno contra la
        inteligencia artificial, y partidas de hasta seis jugadores.

        Limites por modalidad (los aplica el servicio; fuera de ellos, `400`
        con el campo y el limite superado):

        | Modalidad      | maximoParticipantes | heroesIA                 | tamanoEquipo |
        |----------------|---------------------|--------------------------|--------------|
        | UNO_CONTRA_UNO | exactamente 2       | 0                        | no admite    |
        | CONTRA_IA      | exactamente 2       | exactamente 1 (implicito)| no admite    |
        | HASTA_SEIS     | 2 a 6               | 0 a maximoParticipantes-1| 1 a 3        |
      enum: [UNO_CONTRA_UNO, CONTRA_IA, HASTA_SEIS]
...
    ResumenJugador:
      type: object
      required: [id, apodo]
      properties:
        id: { type: string, format: uuid }
        apodo: { type: string }
        avatarUrl: { type: [string, 'null'], format: uri }
        rango: { type: [string, 'null'] }
    Participante:
      type: object
      required: [jugador, heroe, esIA, listo]
      properties:
        jugador: { $ref: '#/components/schemas/ResumenJugador' }
        heroe: { $ref: '#/components/schemas/HeroeEnPartida' }
        esIA:
          type: boolean
          description: Verdadero si lo controla la inteligencia artificial (RF-JUE-001, RF-JUE-004).
        listo: { type: boolean }
        equipo:
          type: [integer, 'null']
          description: Numero de equipo en modo cooperativo.
        creditosApostados: { type: integer, minimum: 0 }
    HeroeEnPartida:
      type: object
      description: |
        Datos del heroe que la sala y el combate necesitan mostrar. El detalle
        completo del heroe pertenece al modulo de contenido: aqui solo viaja
        lo que la vista de batalla pinta.
      required: [id, nombre, vidaActual, vidaMaxima]
      properties:
        id:
          type: string
          minLength: 1
          description: |
            Identificador del heroe en el inventario de su dueno. Es una cadena
            opaca, no un UUID: `ElementoInventario.id` de
            `contracts/openapi/inventario.yaml` lo declara `type: string` y este
            servicio no lo interpreta, solo lo devuelve tal cual lo recibe.
            Exigir aqui formato uuid rompia la verificacion de HU-SAL-003 con
            inventarios perfectamente validos.
        nombre: { type: string }
        retratoUrl: { type: [string, 'null'], format: uri }
        nivel: { type: [integer, 'null'] }
        vidaActual:
          type: integer
          minimum: 0
          description: |
            Vida actual. La barra muestra SIEMPRE el valor numerico junto al
            color (RF-JUE-009): el color nunca es el unico indicador.
        vidaMaxima:
          type: integer
          minimum: 1
          description: |
            Vida maxima. El porcentaje se calcula en el cliente. Los umbrales
            de color son 60 % y 40 % (RF-JUE-009) y viven en el sistema de
            diseno como `--vida-umbral-alto` y `--vida-umbral-medio`; el
            servidor NO envia color.
        efectosActivos:
          type: array
          description: Efectos en curso. Cada uno se representa con icono (RF-JUE-017).
          items:
            type: object
            required: [codigo, nombre]
            properties:
              codigo: { type: string }
              nombre: { type: string }
              turnosRestantes: { type: [integer, 'null'] }
              tipo:
                type: [string, 'null']
                description: Tipo de efecto del motor de combate (BONO_DEFENSA, DANO_POR_TURNO...), 1.7.0.
              valor: { type: [integer, 'null'] }
        prototipo:
          type: [string, 'null']
          description: |
            (1.7.0) Prototipo del catalogo de heroes («Guerrero Tanque»): el
            tipo de heroe, que fija sus acciones y su fila de la Tabla 21. No es
            el nombre propio.
        poderActual:
          type: [integer, 'null']
          minimum: 0
          description: (1.7.0) Poder del heroe ahora (§6.1.1). Nulo antes del combate.
        poderMaximo:
          type: [integer, 'null']
          minimum: 0
        recargas:
          type: object
          additionalProperties: { type: integer, minimum: 1 }
          description: (1.7.0) Por accion en carga, cuantos turnos propios le faltan (§6.1.2).
        acciones:
          type: array
          description: |
            (1.7.0) Lo que este heroe puede jugar AHORA y por que no lo demas,
            calculado por el motor de combate: la interfaz no reimplementa la
            regla. El `codigo` es lo que se manda como `codigoAccion` por el
            canal (`contracts/websocket/salas-partidas.yaml`).
          items:
            $ref: '#/components/schemas/EstadoDeAccion'
    EstadoDeAccion:
      type: object
      description: (1.7.0) Una accion jugable y su disponibilidad, tal como la calcula el motor de combate.
      required: [codigo, nombre, disponible]
      properties:
        codigo: { type: string, example: Golpe con escudo }
        nombre: { type: string, example: Golpe con escudo }
        tipo:
          type: [string, 'null']
          enum: [ATAQUE, DEFENSA, SANACION, SANACION_GRUPAL, REANIMACION, APOYO, null]
        esEpica: { type: boolean }
        costoPoder: { type: [integer, 'null'], description: Nulo si cuesta todo el poder }
        turnosDeCarga: { type: [integer, 'null'] }
        nivelRequerido: { type: [integer, 'null'] }
        disponible: { type: boolean }
        motivo:
          type: [string, 'null']
          description: Por que no esta disponible, apto para el jugador.
    ResumenDePartida:
      type: object
      description: (1.7.0) Una linea del historial del jugador.
      required: [id, idSala, estado, iniciadaEn]
      properties:
        id: { type: string, format: uuid }
        idSala: { type: string, format: uuid }
        modalidad:
          oneOf:
            - $ref: '#/components/schemas/Modalidad'
            - type: 'null'
        estado:
          type: string
          enum: [EN_CURSO, FINALIZADA]
        resultado:
          type: [string, 'null']
          enum: [VICTORIA, DERROTA, EMPATE, null]
          description: Desde el punto de vista de quien consulta; nulo en curso.
        heroe:
          type: [string, 'null']
          description: Nombre del heroe con el que jugo.
        participantes: { type: integer, minimum: 1 }
        iniciadaEn: { type: string, format: date-time }
        finalizadaEn: { type: [string, 'null'], format: date-time }
    PaginaDePartidas:
      type: object
      required: [contenido, pagina, tamano, totalElementos, totalPaginas]
      properties:
        contenido:
          type: array
          items: { $ref: '#/components/schemas/ResumenDePartida' }
        pagina: { type: integer }
        tamano: { type: integer }
        totalElementos: { type: integer }
        totalPaginas: { type: integer }
...
    Partida:
      type: object
      description: |
        Estado del combate (RF-JUE-017). `participantes` va en el ORDEN DE LOS
        TURNOS, sorteado al empezar entre todos (§6.1.3, 1.7.0).
      required: [id, idSala, estado, participantes, turnoActual]
      properties:
        id: { type: string, format: uuid }
        idSala: { type: string, format: uuid }
        estado:
          type: string
          enum: [EN_CURSO, FINALIZADA]
        participantes:
          type: array
          items: { $ref: '#/components/schemas/Participante' }
        turnoActual:
          type: object
          required: [idJugador, numeroTurno]
          properties:
            idJugador: { type: string, format: uuid }
            numeroTurno: { type: integer, minimum: 1 }
            segundosRestantes:
              type: [integer, 'null']
              description: |
                Nulo si no hay tiempo por turno: el documento no fija uno y el
                parametro `salas.partidas.segundos-por-turno` nace sin valor
                (D-B7-14). Con valor, al agotarse el turno pasa sin accion.
        recompensaEnJuego: { type: integer, minimum: 0 }
        iniciadaEn: { type: string, format: date-time }
        resultado:
          type: [string, 'null']
          enum: [GANADOR, EMPATE, null]
          description: |
            (1.7.0) Nulo en curso. `GANADOR` cuando queda en pie un solo heroe
            o un solo equipo (§6.1.3); `EMPATE` cuando no queda nadie (un
            reflejo o un sangrado pueden tumbar a los dos a la vez).
        ganadores:
          type: array
          items: { type: string, format: uuid }
          description: |
            (1.7.0) Quien gano: el heroe en pie o, en combate por equipos,
            TODOS los integrantes del equipo ganador, tambien los caidos
            (D-B7-15). Vacio en curso o en empate.
        equipoGanador:
          type: [integer, 'null']
          description: (1.7.0) Solo en combate por equipos.
        finalizadaEn:
          type: [string, 'null']
          format: date-time
  securitySchemes:
    keycloak:
      type: openIdConnect
      openIdConnectUrl: /realms/nexus-battles/.well-known/openid-configuration
      description: Identidad emitida por Keycloak 26.
security:
  - keycloak: []